requirements.txt
django
djangorestframework
googlemaps
django-multiselectfield
requests
dj-database-url
gunicorn
psycopg2
whitenoise
django-widget-tweaks
